requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.10
django-cors-headers==4.3.1
django-polymorphic==3.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.32.2
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
